packages/issue-sheet/src/requirements.txt
cx_Freeze==7.1.1
pandas==2.0.3
Pillow==8.4.0
Pillow==9.1.1
Pillow==10.3.0
PyYAML==6.0.1
reportlab==4.1.0
XlsxWriter==1.1.8
xlwings==0.31.4
